+++ package.json
@@ -16,8 +16,8 @@
   },
   "scripts": {
     "start": "npm run start:dev",
-    "start:prod": "node dist/main.js",
-    "start:dev": "node --no-warnings=ExperimentalWarning --loader ts-node/esm/transpile-only src/main.ts",
+    "start:prod": "node --es-module-specifier-resolution=node dist/main.js",
+    "start:dev": "node --no-warnings=ExperimentalWarning --es-module-specifier-resolution=node --loader ts-node/esm/transpile-only src/main.ts",
     "build": "tsc"
   },
   "author": "It's not you it's me",
