+++ src/test.ts
@@ -1,0 +1,3 @@
+export function test() {
+  console.log('test');
+}
